… docs/PERFORMANCE_PLAYBOOK.md
… package.json
… reports/web-vitals-2025-06-22T17-37-53-756Z.html
… scripts/__tests__/api-performance-monitor.test.js
… scripts/__tests__/memory-profiler.test.js
… scripts/__tests__/runtime-performance-monitor.test.js
… scripts/__tests__/web-vitals-analyzer.test.js
… scripts/bundle-analyzer.js
… scripts/memory-profiler.js
… scripts/performance-audit.js
… scripts/runtime-performance-monitor.js
… scripts/web-vitals-analyzer.js