## Summary
Implements comprehensive performance monitoring system with bi-weekly review process as requested in #33.

## Changes
- ✅ Created detailed performance monitoring playbook (docs/PERFORMANCE_PLAYBOOK.md)
- ✅ Enhanced Lighthouse audit runner with desktop/mobile support
- ✅ Implemented Web Vitals analyzer for Core Web Vitals tracking
- ✅ Built API performance monitoring with endpoint analysis
- ✅ Created bundle size analyzer with historical comparison
- ✅ Implemented runtime performance monitor using Puppeteer
- ✅ Added memory profiler for leak detection
- ✅ Created master performance audit script to orchestrate all checks
- ✅ Added comprehensive test coverage for all monitoring scripts
- ✅ Configured npm scripts for easy execution

## Performance Targets
- **LCP**: < 2.5s (critical: 4s)
- **FID**: < 100ms (critical: 300ms)
- **CLS**: < 0.1 (critical: 0.25)
- **Bundle Size**: < 200KB (critical: 350KB)
- **API Response**: < 500ms p95 (critical: 1000ms)

## Usage
Run the complete performance audit:
```bash
npm run perf:audit
```

Or run individual checks:
```bash
npm run perf:lighthouse    # Lighthouse audit
npm run perf:bundle       # Bundle size analysis
npm run perf:vitals       # Web Vitals analysis
npm run perf:api          # API performance
npm run perf:runtime      # Runtime performance
npm run perf:memory       # Memory profiling
```

## Test Plan
- [x] All monitoring scripts have comprehensive test coverage
- [x] Scripts are executable and properly configured
- [ ] Manual testing of each performance monitoring tool
- [ ] Verify reports are generated correctly
- [ ] Test on actual deployment

## Notes
- Added puppeteer as dev dependency for runtime/memory profiling
- Reports are saved to `/reports` directory with timestamps
- Historical data is tracked for trend analysis
- Existing tests have some failures unrelated to this PR (pre-existing issues)

Closes #33

🤖 Generated with [Claude Code](https://claude.ai/code)